+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 ### Added
 - Added support for caret position in arguments
+- Added brace matching support.
 
 ### Fixed
 - Fixed argument with argument name not being parsed correctly